add local/p2p/tag push option in notify page

+++ io/Store.js
@@ -3,13 +3,16 @@ import {
   Platform,
 } from 'react-native';
 import SharedPreferences from 'react-native-shared-preferences'
+//import UserDefaults from 'react-native-user-defaults'
 
 var deviceStorage = {
         API_LIST:  "api_list",
         FEED_LIST: "feed_list",
         PUSH_CLICKED: "push_clicked",
-        PUSH_RECEIVED: "pushes_received",
-        P2P_PUSH_LIST: "push_list:r",
+        //PUSH_RECEIVED: "pushes_received",
+        LOCAL_PUSH_LIST: "push_list:local",
+        P2P_PUSH_LIST: "push_list:p2p",
+        TAG_PUSH_LIST: "push_list:tag",
         MAP_LIST:  "map_list",
         PLACE_LIST:"place_list",
         SETTINGS:  "settings",
@@ -26,27 +29,29 @@ var deviceStorage = {
         confTab: 'ion-md-settings',
 
         getShared(key,cbk){
-          if(Platform.OS==='android')
-            SharedPreferences.getItem(key, (value)=>cbk(value))
+          if(Platform.OS==='android') SharedPreferences.getItem(key, (value)=>cbk(value))
+          else if(Platform.OS==='ios') this.get(key).then((value=>cbk(value)))
+          //else if(Platform.OS==='ios') UserDefaults.get(key).then(value=>cbk(value))
         },
         setShared(key,value){
-          if(Platform.OS==='android')
-            SharedPreferences.setItem(key, value)
+          if(Platform.OS==='android') SharedPreferences.setItem(key, value)
+          //else if(Platform.OS==='ios') UserDefaults.set(key, value)
         },
         deleteShared(key){
-          if(Platform.OS==='android')
-            SharedPreferences.deleteItem(key)
+          if(Platform.OS==='android') SharedPreferences.deleteItem(key)
+          //else if(Platform.OS==='ios') UserDefaults.remove(key)
         },
         clearShared(){
-          if(Platform.OS==='android')
-            SharedPreferences.clear()
+          if(Platform.OS==='android') SharedPreferences.clear()
+          //else if(Platform.OS==='ios') UserDefaults.empty()
         },
-        readPushShared(kv) {
+        readPushShared(type,kv) {
             let self = this
-            this.getShared(this.PUSH_RECEIVED,function(str){
+            let fskey = type==='p2p'?this.P2P_PUSH_LIST:this.TAG_PUSH_LIST
+            this.getShared(fskey,function(str){
                 let array = JSON.parse(str)
                 self.addArrayElementShared(array,kv,{read:1})
-                self.setShared(self.PUSH_RECEIVED,JSON.stringify(array));
+                self.setShared(fskey,JSON.stringify(array));
                 //alert('rm:'+id+'\nlist='+JSON.stringify(array))
             })
         },
@@ -63,12 +68,13 @@ var deviceStorage = {
                 //alert('arr='+JSON.stringify(arr))
             }else{}
         },
-        deletePushShared(kv) {
+        deletePushShared(type,kv) {
             let self = this
-            this.getShared(this.PUSH_RECEIVED,function(str){
+            let fskey = type==='p2p'?this.P2P_PUSH_LIST:this.TAG_PUSH_LIST
+            this.getShared(fskey,function(str){
                 let array = JSON.parse(str)
                 self.deleteArrayElementShared(array,kv)
-                self.setShared(self.PUSH_RECEIVED,JSON.stringify(array));
+                self.setShared(fskey,JSON.stringify(array));
                 //alert('rm:'+id+'\nlist='+JSON.stringify(array))
             })
         },
